Matrix File Reading Continue 2
Matrix File Reading Continue:
- File reading partially implemented;

+++ CSC475Lab3/src/csc475lab3/Matrix.java
@@ -45,7 +45,8 @@ public class Matrix {
         StringBuilder builder = new StringBuilder();
         for(float matrixRow[] : this.matrix) {
             for(float i : matrixRow) {
-                builder.append(String.format("%.0f ", i));
+                //builder.append(String.format("%.0f ", i));
+                builder.append(String.format("%6d ", (long)i));
             }
              builder.append("\n");
         }
